requirements.txt
absl-py==1.4.0
altair==4.2.0
ansi2html==1.9.0rc1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.0
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blis==0.7.9
bokeh==3.0.3
branca==0.6.0
cachetools==5.2.0
cargo==0.3
catalogue==2.0.8
category-encoders==2.6.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.0
colorama==0.4.6
colorlover==0.3.0
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.6
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==3.0.0a11
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datamallet==0.22.0
datascience==0.17.5
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastai==2.7.10
fastcore==1.5.28
fastdownload==0.0.7
fastjsonschema==2.16.2
fastprogress==1.0.3
Fiona==1.8.22
fire==0.4.0
Flask==2.1.3
Flask-Cors==3.0.10
flatbuffers==23.1.21
folium==0.13.0
fonttools==4.38.0
fqdn==1.5.1
gast==0.4.0
geojson==2.5.0
geopandas==0.11.1
gitdb==4.0.10
GitPython==3.1.29
gmaps==0.9.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
html5lib==1.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.10.1
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kaggle==1.5.12
kaleido==0.2.1
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lightgbm==3.3.4
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.0.0
multimethod==1.9.1
multitasking==0.0.11
munch==2.5.0
murmurhash==1.0.9
music21==8.1.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.8.5
packaging==21.3
pandas==1.5.2
pandas-profiling==3.6.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
plotly==5.11.0
plotly-express==0.4.1
plotly-resampler==0.8.4rc1
pmdarima==2.0.2
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==3.0.0rc8
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
pygame==2.1.2
Pygments==2.13.0
Pympler==1.0.1
pyod==1.0.7
pyparsing==3.0.9
pyphen==0.13.2
pyproj==3.4.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
schemdraw==0.15
scikit-learn >=0.20.0<1.2.0
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.1
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.14.0
setproctitle==1.3.2
Shapely==1.8.5.post1
six==1.16.0
sktime==0.15.1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.17.0
streamlit-option-menu==0.3.2
tangled-up-in-unicode==0.2.0
tbats==1.1.2
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.1.1
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.7
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
trace-updater==0.0.9
traitlets==5.6.0
typeguard==2.13.3
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
unzip==1.0.0
uri-template==1.2.0
urllib3==1.26.13
validators==0.20.0
visions==0.7.5
wandb==0.13.9
wasabi==1.1.1
watchdog==2.1.9
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==4.0.4
wordcloud==1.8.2.2
wrapt==1.15.0rc1
xgboost==1.7.3
XlsxWriter==3.0.3
xyzservices==2022.9.0
yarg==0.1.9
yellowbrick==1.5
yfinance==0.1.74
zipp==3.11.0